ref(event-details): Convert EventErrors to an FC and use react-query
